# apps/api/src/handlers/index.ts
import { FastifyBaseLogger, FastifyInstance, FastifyReply, FastifyRequest, FastifySchema, FastifyTypeProviderDefault, RawServerDefault } from "fastify";
import { ResolveFastifyRequestType } from "fastify/types/type-provider";
import { IncomingMessage, ServerResponse } from "node:http";
import type { Handlers } from "../generated/openapi/handlers";
import { components } from "../generated/openapi/types";
import { login } from "./auth";
import { getHealth } from "./health";

export const handlers: Handlers = {
    getHealth: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: undefined; Querystring: undefined; Body: unknown; Reply: { status: string; database: string; }; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: undefined; Querystring: undefined; Body: unknown; Reply: { status: string; database: string; }; Headers: undefined; }>>, reply: FastifyReply<{ Params: undefined; Querystring: undefined; Body: unknown; Reply: { status: string; database: string; }; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, { status: string; database: string; }>): void | { status: string; database: string; } | Promise<void | { status: string; database: string; }> {
        return getHealth();
    },
    login: async function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: undefined; Querystring: undefined; Body: { email: string; password: string; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: undefined; Querystring: undefined; Body: { email: string; password: string; }; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: undefined; Querystring: undefined; Body: { email: string; password: string; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): Promise<unknown> {
        const isSuccess = await login(request.body.email, request.body.password);
        if (isSuccess) {
            return reply.status(200).send({ message: "Login successful" });
        } else {
            return reply.status(401).send({ message: "Invalid email or password" });
        }
    },
    getMe: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: undefined; Querystring: undefined; Body: unknown; Reply: { id: string; name: string; email: string; }; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: undefined; Querystring: undefined; Body: unknown; Reply: { id: string; name: string; email: string; }; Headers: undefined; }>>, reply: FastifyReply<{ Params: undefined; Querystring: undefined; Body: unknown; Reply: { id: string; name: string; email: string; }; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, { id: string; name: string; email: string; }>): void | { id: string; name: string; email: string; } | Promise<void | { id: string; name: string; email: string; }> {
        throw new Error("Function not implemented.");
    },
    getDashboard: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: undefined; Querystring: undefined; Body: unknown; Reply: { progress: components["schemas"]["Progress"]; groups: { id: string; name: string; purchasedItems: number; totalItems: number; percentage: number; }[]; }; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: undefined; Querystring: undefined; Body: unknown; Reply: { progress: components["schemas"]["Progress"]; groups: { id: string; name: string; purchasedItems: number; totalItems: number; percentage: number; }[]; }; Headers: undefined; }>>, reply: FastifyReply<{ Params: undefined; Querystring: undefined; Body: unknown; Reply: { progress: components["schemas"]["Progress"]; groups: { id: string; name: string; purchasedItems: number; totalItems: number; percentage: number; }[]; }; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, { progress: components["schemas"]["Progress"]; groups: { id: string; name: string; purchasedItems: number; totalItems: number; percentage: number; }[]; }>): void | { progress: components["schemas"]["Progress"]; groups: { id: string; name: string; purchasedItems: number; totalItems: number; percentage: number; }[]; } | Promise<void | { progress: components["schemas"]["Progress"]; groups: { id: string; name: string; purchasedItems: number; totalItems: number; percentage: number; }[]; }> {
        throw new Error("Function not implemented.");
    },
    getGroups: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: undefined; Querystring: undefined; Body: unknown; Reply: { id: string; name: string; }[]; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: undefined; Querystring: undefined; Body: unknown; Reply: { id: string; name: string; }[]; Headers: undefined; }>>, reply: FastifyReply<{ Params: undefined; Querystring: undefined; Body: unknown; Reply: { id: string; name: string; }[]; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, { id: string; name: string; }[]>): void | { id: string; name: string; }[] | Promise<void | { id: string; name: string; }[]> {
        throw new Error("Function not implemented.");
    },
    createGroup: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: undefined; Querystring: undefined; Body: { name: string; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: undefined; Querystring: undefined; Body: { name: string; }; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: undefined; Querystring: undefined; Body: { name: string; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): unknown {
        throw new Error("Function not implemented.");
    },
    getGroup: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): unknown {
        throw new Error("Function not implemented.");
    },
    updateGroup: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: { name: string; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: { name: string; }; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: { name: string; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): unknown {
        throw new Error("Function not implemented.");
    },
    deleteGroup: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): unknown {
        throw new Error("Function not implemented.");
    },
    createItem: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: { name: string; description?: string | null; url?: string | null; plans: components["schemas"]["ItemPlan"][]; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: { name: string; description?: string | null; url?: string | null; plans: components["schemas"]["ItemPlan"][]; }; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: { groupId: components["parameters"]["GroupId"]; }; Querystring: undefined; Body: { name: string; description?: string | null; url?: string | null; plans: components["schemas"]["ItemPlan"][]; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): unknown {
        throw new Error("Function not implemented.");
    },
    getItem: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): unknown {
        throw new Error("Function not implemented.");
    },
    updateItem: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: { name: string; description?: string | null; url?: string | null; plans: components["schemas"]["ItemPlan"][]; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: { name: string; description?: string | null; url?: string | null; plans: components["schemas"]["ItemPlan"][]; }; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: { name: string; description?: string | null; url?: string | null; plans: components["schemas"]["ItemPlan"][]; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): unknown {
        throw new Error("Function not implemented.");
    },
    deleteItem: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): unknown {
        throw new Error("Function not implemented.");
    },
    purchaseItem: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: { amount: number; currency: string; purchasedAt: string; memo?: string | null; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: { amount: number; currency: string; purchasedAt: string; memo?: string | null; }; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: { amount: number; currency: string; purchasedAt: string; memo?: string | null; }; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): unknown {
        throw new Error("Function not implemented.");
    },
    cancelPurchase: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }>>, reply: FastifyReply<{ Params: { itemId: components["parameters"]["ItemId"]; }; Querystring: undefined; Body: unknown; Reply: unknown; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, unknown>): unknown {
        throw new Error("Function not implemented.");
    },
    getExchangeRates: function (this: FastifyInstance<RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, FastifyBaseLogger, FastifyTypeProviderDefault>, request: FastifyRequest<{ Params: undefined; Querystring: undefined; Body: unknown; Reply: { base: string; rates: { [key: string]: number; }; updatedAt: string; }; Headers: undefined; }, RawServerDefault, IncomingMessage, FastifySchema, FastifyTypeProviderDefault, unknown, FastifyBaseLogger, ResolveFastifyRequestType<FastifyTypeProviderDefault, FastifySchema, { Params: undefined; Querystring: undefined; Body: unknown; Reply: { base: string; rates: { [key: string]: number; }; updatedAt: string; }; Headers: undefined; }>>, reply: FastifyReply<{ Params: undefined; Querystring: undefined; Body: unknown; Reply: { base: string; rates: { [key: string]: number; }; updatedAt: string; }; Headers: undefined; }, RawServerDefault, IncomingMessage, ServerResponse<IncomingMessage>, unknown, FastifySchema, FastifyTypeProviderDefault, { base: string; rates: { [key: string]: number; }; updatedAt: string; }>): void | { base: string; rates: { [key: string]: number; }; updatedAt: string; } | Promise<void | { base: string; rates: { [key: string]: number; }; updatedAt: string; }> {
        throw new Error("Function not implemented.");
    },
};